refactor: comment 패키지 web Command, Query로 분리
