+++ src/main/java/dandi/dandi/comment/adapter/in/web/CommentUseCaseControllerDocs.java
@@ -2,7 +2,6 @@ package dandi.dandi.comment.adapter.in.web;
 
 import dandi.dandi.advice.ExceptionResponse;
 import dandi.dandi.comment.application.port.in.CommentRegisterCommand;
-import dandi.dandi.comment.application.port.in.CommentResponses;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -10,13 +9,12 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 
 @Tag(name = "댓글")
-public interface CommentControllerDocs {
+public interface CommentUseCaseControllerDocs {
 
     @Operation(summary = "댓글 작성")
     @ApiResponses(value = {
@@ -27,16 +25,6 @@ public interface CommentControllerDocs {
     ResponseEntity<Void> registerComment(@Parameter(hidden = true) Long memberId, @PathVariable Long postId,
                                          @RequestBody CommentRegisterCommand commentRegisterCommand);
 
-    @Operation(summary = "댓글 조회", parameters = {
-            @Parameter(name = "size"), @Parameter(name = "page"),
-            @Parameter(name = "sort"), @Parameter(example = "DESC")})
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "댓글 조회 성공"),
-            @ApiResponse(responseCode = "404", description = "존재하지 않는 게시글에 대한 댓글 조회 요청",
-                    content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
-    })
-    ResponseEntity<CommentResponses> getComments(@Parameter(hidden = true) Long memberId, @PathVariable Long postId,
-                                                 @Parameter(hidden = true) Pageable pageable);
 
     @Operation(summary = "댓글 삭제")
     @ApiResponses(value = {
@@ -49,13 +37,4 @@ public interface CommentControllerDocs {
 
     )
     ResponseEntity<Void> deleteComment(@Parameter(hidden = true) Long memberId, @PathVariable Long commentId);
-
-
-    @Operation(summary = "댓글 신고")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "201", description = "댓글 신고 성공"),
-            @ApiResponse(responseCode = "404", description = "존재하지 않는 댓글"),
-            @ApiResponse(responseCode = "400", description = "이미 신고한 댓글"),
-    })
-    ResponseEntity<Void> reportComment(@Parameter(hidden = true) Long memberId, @PathVariable Long commentId);
 }
